Update space_invader.py, player_die.png, invader_die.png : Player, invader, ufo 사망 애니메이션 추가

+++ space_invader.py
@@ -1,6 +1,5 @@
 import gui_core as gui
 import random
-import time
 
 screen_width = 600
 screen_height = 800
@@ -61,7 +60,7 @@ def initialize(timestamp):
 	player_x = (screen_width / 2) - (w.data.player_width / 2)
 	player_y = w.data.game_over_line_y - (w.data.player_height * 2.5)
 	playerNumber = w.newImage(player_x, player_y, 'player.png', w.data.player_width, w.data.player_height)
-	w.data.objs.append(['player', playerNumber, player_x, player_y, w.data.player_width])
+	w.data.objs.append(['player', playerNumber, player_x, player_y, w.data.player_width, False, 0])
 
 	tNum = w.newText(x_offset + 25, w.data.game_over_line_y + 20, 50, str(w.data.player_life), 'white')
 	w.data.objs.append(['life_text', tNum])
@@ -79,11 +78,11 @@ def initialize(timestamp):
 			pos_y = 382 - (w.data.invader_interval_v * i)
 			number = w.newImage(pos_x, pos_y, w.data.filenames[fileNameIdx][0], w.data.invader_width[fileNameIdx], w.data.invader_height[fileNameIdx])
 			w.data.objs.append([
-				'invader', number, fileNameIdx, 0, pos_x, pos_y, timestamp + (i / 20) + (j / 50), 0, 1, 1
+				'invader', number, fileNameIdx, 0, pos_x, pos_y, timestamp + (i / 20) + (j / 50), 0, 1, 1, False, 0
 			])
 
 	ufoNum = w.newImage(w.data.ufo_x, w.data.ufo_y, 'ufo.png', w.data.ufo_width, w.data.ufo_height)
-	w.data.objs.append(['ufo', ufoNum, w.data.ufo_x, w.data.ufo_y, timestamp])
+	w.data.objs.append(['ufo', ufoNum, w.data.ufo_x, w.data.ufo_y, timestamp, False, 0])
 	w.newRectangle(x_offset, w.data.game_over_line_y, screen_width - (x_offset * 2), 3, 'green')
 
 def update(timestamp):
@@ -106,34 +105,40 @@ def update(timestamp):
 
 	if w.data.isGameOver == False:
 		w.data.invader_count = 0
-
 		for obj in w.data.objs:
 			if obj[0] == 'player':
-				if w.keys['Left']:
-					if obj[2] > 8 + x_offset:
-						obj[2] -= 3
-					w.moveObject(obj[1], obj[2], obj[3])
-				if w.keys['Right']:
-					if obj[2] < screen_width - x_offset - w.data.player_width - 8:
-						obj[2] += 3
-					w.moveObject(obj[1], obj[2], obj[3])
-				if w.keys['space'] and timestamp - w.data.last_missile_time > 0.5:
-					mNum = w.newImage(obj[2], obj[3], w.data.missilefiles[0], w.data.missile_width, w.data.missile_height)
-					w.data.objs.append(['missile', mNum, obj[2], obj[3], 0, timestamp])
-					w.data.last_missile_time = timestamp
+				if obj[5] and timestamp - obj[6] > 0.3:
+					w.setImage(obj[1], 'player.png', w.data.player_width, w.data.player_height)
+					w.data.player_life -= 1
+					if w.data.player_life <= 0:
+						w.deleteObject(obj[1])
+						w.data.objs.remove(obj)
+						setGameOver()
+					obj[5] = False
+				if obj[5] == False:
+					if w.keys['Left']:
+						if obj[2] > 8 + x_offset:
+							obj[2] -= 3
+						w.moveObject(obj[1], obj[2], obj[3])
+					if w.keys['Right']:
+						if obj[2] < screen_width - x_offset - w.data.player_width - 8:
+							obj[2] += 3
+						w.moveObject(obj[1], obj[2], obj[3])
+					if w.keys['space'] and timestamp - w.data.last_missile_time > 0.5:
+						mNum = w.newImage(obj[2], obj[3], w.data.missilefiles[0], w.data.missile_width, w.data.missile_height)
+						w.data.objs.append(['missile', mNum, obj[2], obj[3], 0, timestamp])
+						w.data.last_missile_time = timestamp
 				
-				for missile in [m for m in w.data.objs if m[0] == 'invader_missile']:
-					m_x = missile[2] + (w.data.invader_missile_width / 2)
-					m_y = missile[3]
-					if m_x > obj[2] and m_x < obj[2] + w.data.player_width and m_y > obj[3] and m_y < obj[3] + w.data.player_height:
-						w.data.player_life -= 1
-						if w.data.player_life <= 0:
-							w.deleteObject(obj[1])
-							w.data.objs.remove(obj)
-							setGameOver()
-						w.deleteObject(missile[1])
-						w.data.objs.remove(missile)
-						break
+					for missile in [m for m in w.data.objs if m[0] == 'invader_missile']:
+						m_x = missile[2] + (w.data.invader_missile_width / 2)
+						m_y = missile[3]
+						if m_x > obj[2] and m_x < obj[2] + w.data.player_width and m_y > obj[3] and m_y < obj[3] + w.data.player_height:
+							w.deleteObject(missile[1])
+							w.data.objs.remove(missile)
+							w.setImage(obj[1], 'player_die.png', 30, 16)
+							obj[5] = True
+							obj[6] = timestamp
+							break
 
 			elif obj[0] == 'missile':
 				'''
@@ -153,20 +158,39 @@ def update(timestamp):
 					obj[5] = timestamp
 
 			elif obj[0] == 'ufo':
-				if timestamp - obj[4] > 15:
+				'''
+				ufo[0]: objName
+				ufo[1]: number
+				ufo[2]: pos_x
+				ufo[3]: pos_y
+				ufo[4]: resetTime
+				ufo[5]: isDead
+				ufo[6]: deadTime
+				'''
+				if obj[5] and timestamp - obj[6] > 0.3:	
+					ufo_score = random.randint(50, 300)
+					w.data.score += ufo_score
+					obj[2] = w.data.ufo_x
+					obj[4] = timestamp
+					obj[5] = False
+
+				if obj[5] == False and timestamp - obj[4] > 3:
 					obj[2] += 4
+
 				w.moveObject(obj[1], obj[2], obj[3])
+
 				if obj[2] > screen_width:
 					obj[2] = w.data.ufo_x
 					obj[4] = timestamp
+					obj[5] = False
+
 				for missile in [m for m in w.data.objs if m[0] == 'missile']:
 					m_x = missile[2] + (w.data.missile_width / 2)
 					m_y = missile[3]
 					if m_x > obj[2] and m_x < obj[2] + w.data.ufo_width and m_y > obj[3] and m_y < obj[3] + w.data.ufo_height:
-						ufo_score = random.randint(50, 300)
-						w.data.score += ufo_score
-						w.deleteObject(obj[1])
-						w.data.objs.remove(obj)
+						w.setImage(obj[1], 'invader_die.png', 52, 32)
+						obj[5] = True
+						obj[6] = timestamp
 						w.deleteObject(missile[1])
 						w.data.objs.remove(missile)
 						break
@@ -183,11 +207,18 @@ def update(timestamp):
 				invader[7]: moveCnt
 				invader[8]: moveDir
 				invader[9]: timeMod
+				invader[10]: isDead
+				invader[11]: deadTime
 				'''
 				if w.keys['c']:
 					w.deleteObject(obj[1])
 					w.data.objs.remove(obj)
 				
+				if obj[10] and timestamp - obj[11] > 0.2:	
+					w.data.score += 30 - (obj[2] * 10)
+					w.deleteObject(obj[1])
+					w.data.objs.remove(obj)
+
 				if obj[5] + w.data.invader_height[obj[2]] >= w.data.game_over_line_y:
 					setGameOver()
 
@@ -195,9 +226,9 @@ def update(timestamp):
 					m_x = missile[2] + (w.data.missile_width / 2)
 					m_y = missile[3]
 					if m_x > obj[4] and m_x < obj[4] + w.data.invader_width[obj[2]] and m_y > obj[5] and m_y < obj[5] + w.data.invader_height[obj[2]]:
-						w.data.score += 30 - (obj[2] * 10)
-						w.deleteObject(obj[1])
-						w.data.objs.remove(obj)
+						w.setImage(obj[1], 'invader_die.png', 26, 16)
+						obj[10] = True
+						obj[11] = timestamp
 						w.deleteObject(missile[1])
 						w.data.objs.remove(missile)
 						break
@@ -207,7 +238,7 @@ def update(timestamp):
 					w.data.objs.append(['invader_missile', mNum, obj[4], obj[5], 0, timestamp])
 					w.data.last_invader_missile_time = timestamp
 
-				if obj[6] <= timestamp:
+				if obj[10] == False and obj[6] <= timestamp:
 					obj[2] %= 3
 					obj[3] = (obj[3] + 1) % 2
 					obj[4] += 8 * obj[8]
